Fix code splitting issue in application-code-utils.ts
