+++ frontend/app/routes/$lang/_public/status/child.tsx
@@ -26,7 +26,8 @@ import { LoadingButton } from '~/components/loading-button';
 import { getHCaptchaRouteHelpers } from '~/route-helpers/h-captcha-route-helpers.server';
 import { getApplicationStatusService } from '~/services/application-status-service.server';
 import { getLookupService } from '~/services/lookup-service.server';
-import { applicationCodeInputPatternFormat, getContextualAlertType, isValidCodeOrNumber } from '~/utils/application-code-utils';
+import { applicationCodeInputPatternFormat, isValidCodeOrNumber } from '~/utils/application-code-utils';
+import { getContextualAlertType } from '~/utils/application-code-utils.server';
 import { extractDateParts, getAgeFromDateString, isPastDateString, isValidDateString } from '~/utils/date-utils';
 import { featureEnabled, getEnv } from '~/utils/env.server';
 import { useHCaptcha } from '~/utils/hcaptcha-utils';
